SaiSeleniumTests
Compiled and executed in J2EE environment, functional tests for user story 1 and user story 2 added.
 
In this webpage we are using Canvas for drawing the pie chart and so it’s tough to get the data/control over the Pie Chart. But it would have made the job easier if the pie chart controls are developed using svg (HTML5 controls) so that finding these we can to use //* in xpath. In that case we can find the dynamic sections in the pie chart and click each section to find the tool tip value.
 
driver.findElement(By.xpath("//div[contains(@id,'_tooltip')]")); -- after finding this we can click a section by finding it using driver.findElement(By.xpath("//*   [contains(@id,'p_3_1_2_4_')][contains(@class,'3-PresidentPieSection')][@fill='#ffffff']")); using click() function followed by getText() function to get the Tool Tip value.
 
Given a chance of this kind of chart we should be able to automate whether it might be a Bar Chart/Pie Chart using selenium web driver.